move noti to a different file
