+++ src/components/GameSelect.js
@@ -7,6 +7,8 @@ import webRTC from '../api/WebRTC.js';
 import "./GameSelect.css";
 import "../common/Animations.css";
 
+import { openNotification } from "./Notification"
+
 function GameSelect(props) {
     const history = useHistory();
     const location = useLocation();
@@ -14,7 +16,7 @@ function GameSelect(props) {
     useEffect(() => {
         if (!location.username) {
             // window.alert("Please log in before playing a game!");
-            props.openNotification('error', "Login Error", "Please log in before playing a game!");
+            openNotification('error', "Login Error", "Please log in before playing a game!");
             history.push("/");
         }
     }, []);
@@ -23,14 +25,14 @@ function GameSelect(props) {
         signallingServer.findRobot("battle")
             .then((robotName) => {
                 console.log("Initializing peer connection");
-                props.openNotification('info', "", "Looking for robots...");
+                openNotification('info', "", "Looking for robots...");
                 webRTC.initializePeerConnection()
                     .then(() => {
                         console.log("Sending offer to robot: " + robotName);
                         signallingServer.sendOffer(robotName, webRTC.getOffer())
                             .then((answer) => {
                                 console.log("Game can be started");
-                                props.openNotification('success', "", "Game successfully started!");
+                                openNotification('success', "", "Game successfully started!");
                                 webRTC.setAnswer(answer);
                                 signallingServer.startGame();
 
@@ -40,16 +42,16 @@ function GameSelect(props) {
                                 });
                             })
                             .catch((error) => {
-                                props.openNotification('error',"Error", error.message);
+                                openNotification('error',"Error", error.message);
                             });
                     })
                     .catch((error) => {
-                        props.openNotification('error',"Error", error.message);
+                        openNotification('error',"Error", error.message);
                     });
             })
             .catch((error) => {
                 // TODO Handle no robot found
-                props.openNotification('error',"Error", error.message);
+                openNotification('error',"Error", error.message);
             })
     }
